@@ -6,7 +6,8 @@
     "homepage": "http://github.com/pmjones/auto-route",
     "license": "MIT",
     "require": {
-        "php": ">=7.2.0"
+        "php": "^8.0",
+        "psr/log": "^3.0"
     },
     "autoload": {
         "psr-4": {
@@ -19,10 +20,15 @@
         }
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.0",
-        "pds/skeleton": "^1.0"
+        "pds/skeleton": "^1.0",
+        "phpunit/phpunit": "^9.0",
+        "phpstan/phpstan": "^0.12.82"
     },
     "bin": [
         "bin/autoroute-dump.php"
-    ]
+    ],
+    "scripts": {
+        "test": "./vendor/bin/phpunit",
+        "stan": "./vendor/bin/phpstan analyze -c phpstan.neon src"
+    }
 }
