@@ -12,87 +12,17 @@ namespace AutoRoute;
 
 class Actions
 {
-    protected $namespace;
+    protected array $instances = [];
 
-    protected $namespaceLen;
+    protected array $reversals = [];
 
-    protected $instances = [];
-
-    protected $wordSeparator;
-
-    protected $directory;
-
-    protected $method;
-
-    protected $baseUrl;
-
-    protected $baseUrlLen;
-
-    protected $ignoreParams;
-
-    protected $suffix;
-
-    protected $suffixLen;
+    protected Reverser $reverser;
 
     public function __construct(
-        string $namespace,
-        string $directory,
-        string $suffix,
-        string $method,
-        int $ignoreParams,
-        string $baseUrl,
-        string $wordSeparator
+        protected Config $config,
+        protected Reflector $reflector,
     ) {
-        $this->namespace = trim($namespace, '\\') . '\\';
-        $this->namespaceLen = strlen($this->namespace);
-        $this->directory = rtrim($directory, DIRECTORY_SEPARATOR);
-        $this->suffix = $suffix;
-        $this->suffixLen = strlen($suffix);
-        $this->method = $method;
-        $this->ignoreParams = $ignoreParams;
-        $this->baseUrl = trim($baseUrl, '/');
-        $this->baseUrlLen = strlen($this->baseUrl);
-        $this->wordSeparator = $wordSeparator;
-    }
-
-    public function getNamespace() : string
-    {
-        return $this->namespace;
-    }
-
-    public function getMethod() : string
-    {
-        return $this->method;
-    }
-
-    public function getDirectory() : string
-    {
-        return $this->directory;
-    }
-
-    public function segmentToNamespace(string $segment) : string
-    {
-        $segment = trim($segment);
-
-        if ($segment === '') {
-            throw new InvalidNamespace("Cannot convert empty segment to namespace part");
-        }
-
-        return str_replace(
-            $this->wordSeparator,
-            '',
-            ucwords($segment, $this->wordSeparator)
-        );
-    }
-
-    public function namespaceToSegment(string $part) : string
-    {
-        $part = trim($part);
-        return strtolower(preg_replace(
-            '/([a-z])([A-Z])/',
-            "\$1{$this->wordSeparator}\$2",
-            $part
-        ));
+        $this->reverser = new Reverser($this->config, $this);
     }
 
     public function getAction(string $class) : Action
@@ -106,70 +36,56 @@ class Actions
 
     protected function newAction(string $class) : Action
     {
-        $ns = substr($class, 0, $this->namespaceLen);
-        if ($ns !== $this->namespace) {
-            throw new InvalidNamespace("Expected namespace {$this->namespace}, actually $class");
+        $ns = substr($class, 0, $this->config->namespaceLen);
+
+        if ($ns !== $this->config->namespace) {
+            throw new Exception\InvalidNamespace("Expected namespace {$this->config->namespace}, actually {$class}");
         }
 
-        if (! class_exists($class)) {
-            throw new NotFound("Expected class $class, actually not found");
+        if (! $this->reflector->classExists($class)) {
+            throw new Exception\NotFound("Expected class {$class}, actually not found");
+        }
+
+        $parameters = $this->reflector->getActionParameters($class);
+
+        for ($i = 0; $i < $this->config->ignoreParams; $i ++) {
+            array_shift($parameters);
+        }
+
+        $requiredParameters = [];
+        $optionalParameters = [];
+
+        foreach ($parameters as $i => $rp) {
+            if ($rp->isOptional()) {
+                $optionalParameters[$i] = $rp;
+            } else {
+                $requiredParameters[$i] = $rp;
+            }
         }
 
         return new Action(
-            $this->namespace,
             $class,
-            $this->suffix,
-            $this->method,
-            $this->ignoreParams,
-            $this->baseUrl
+            $requiredParameters,
+            $optionalParameters
         );
     }
 
-    public function generate(string $class) : array
+    public function getReverse(string $class) : Reverse
     {
-        return $this->getAction($class)->generate($this);
-    }
-
-    public function dump(string $class) : array
-    {
-        return $this->getAction($class)->dump($this);
-    }
-
-    public function isSubNamespace(string $subns) : bool
-    {
-        if (substr($subns, -2) == '..') {
-            throw new InvalidNamespace("Directory dots not allowed in segments");
+        if (! isset($this->reversals[$class])) {
+            $this->reversals[$class] = $this->reverser->reverse($class);
         }
 
-        $dir = $this->directory . str_replace('\\', DIRECTORY_SEPARATOR, $subns);
-        return is_dir($dir);
+        return $this->reversals[$class];
     }
 
-    public function getSegments(string $path) : array
-    {
-        $path = trim($path, '/');
-        $base = substr($path, 0, $this->baseUrlLen);
-        if ($base !== $this->baseUrl) {
-            throw new NotFound("Expected base URL /$this->baseUrl, actually /$base");
-        }
-
-        $segments = [];
-
-        $path = trim(substr($path, $this->baseUrlLen), '/');
-        if (! empty($path)) {
-            $segments = explode('/', $path);
-        }
-
-        return $segments;
-    }
-
-    public function actionExists(
+    public function getClass(
         string $verb,
         string $subNamespace,
         string $append = ''
     ) : ?string
     {
-        $base = rtrim($this->namespace, '\\')
+        $base = rtrim($this->config->namespace, '\\')
             . $subNamespace
             . '\\';
 
@@ -177,10 +93,19 @@ class Actions
             $base .= $append . '\\';
         }
 
-        $ending = str_replace('\\', '', $subNamespace . $append) . $this->suffix;
-        $class = $base . $verb . $ending;
+        $ending = str_replace('\\', '', $subNamespace . $append) . $this->config->suffix;
+        return $base . $verb . $ending;
+    }
 
-        if (class_exists($class)) {
+    public function hasAction(
+        string $verb,
+        string $subNamespace,
+        string $append = ''
+    ) : ?string
+    {
+        $class = $this->getClass($verb, $subNamespace, $append);
+
+        if ($this->reflector->classExists($class)) {
             return $class;
         }
 
@@ -188,27 +113,7 @@ class Actions
             return null;
         }
 
-        $getClass = $base . 'Get' . $ending;
-        return class_exists($getClass) ? $getClass : null;
-    }
-
-    public function fileToClass(string $file) : ?string
-    {
-        $file = str_replace($this->directory . DIRECTORY_SEPARATOR, '', substr($file, 0, -4));
-        $parts = explode(DIRECTORY_SEPARATOR, $file);
-
-        $last = array_pop($parts);
-        $core = implode('', $parts);
-        $verb = substr($last, 0, strlen($last) - strlen($core) - $this->suffixLen);
-        if ($verb === '') {
-            return null;
-        }
-
-        $subNamespace = '';
-        if (! empty($parts)) {
-            $subNamespace = '\\' . implode('\\', $parts);
-        }
-
-        return $this->actionExists($verb, $subNamespace);
+        $class = $this->getClass('Get', $subNamespace, $append);
+        return $this->reflector->classExists($class) ? $class : null;
     }
 }
