@@ -10,27 +10,24 @@ declare(strict_types=1);
 
 namespace AutoRoute;
 
-class Action
+class NullAction extends Action
 {
-    public function __construct(
-        protected string $class,
-        protected array $requiredParameters,
-        protected array $optionalParameters,
-    ) {
+    public function __construct()
+    {
     }
 
     public function getClass() : string
     {
-        return $this->class;
+        return '';
     }
 
     public function getRequiredParameters(int $offset = 0) : array
     {
-        return array_slice($this->requiredParameters, $offset, null, true);
+        return [];
     }
 
     public function getOptionalParameters() : array
     {
-        return $this->optionalParameters;
+        return [];
     }
 }
