@@ -10,17 +10,23 @@ declare(strict_types=1);
 
 namespace AutoRoute;
 
+use ReflectionParameter;
+
 /**
  * @property-read string $class
- * @property-read string $method
- * @property-read array $arguments
+ * @property-read string $verb
+ * @property-read string $path
+ * @property-read array $parameters
+ * @property-read int $requiredParametersTotal
  */
-class Route
+class Reverse
 {
     public function __construct(
         protected string $class,
-        protected string $method,
-        protected array $arguments,
+        protected string $verb,
+        protected string $path,
+        protected array $parameters,
+        protected int $requiredParametersTotal,
     ) {
     }
 
